feat(api-reference): add sdk registration url generation
